Extracting `Resource` to a mixin module `Entity`
With this refactor, you can create new Resource types by including the
mixin module `Entity`.

   class MyResource
     include ActiveTriples::Entity

     ...
   end

You can continue to subclass `Resource`, if desired.
