+++ lib/active_triples/identifiable.rb
@@ -7,12 +7,12 @@ module ActiveTriples::Identifiable
   delegate :rdf_subject, :type, to: :resource
 
   ##
-  # @return [ActiveTriples::Resource] a resource that contains this object's 
+  # @return [ActiveTriples::Resource] a resource that contains this object's
   # graph.
   def resource
    @resource ||= resource_class.new(to_uri)
   end
-  
+
   def parent
     @parent ||= resource.parent
   end
@@ -21,7 +21,7 @@ module ActiveTriples::Identifiable
     @parent = val
   end
 
-  ## 
+  ##
   # @return [String] a uri or slug
   def to_uri
     return id if respond_to? :id and !resource_class.base_uri.nil?
