+++ spec/active_triples/resource_spec.rb
@@ -1,13 +1,15 @@
 require 'spec_helper'
 
 describe ActiveTriples::Resource do
-  it_behaves_like 'an ActiveModel'
+  # it_behaves_like 'an ActiveModel'
   before do
-    class DummyLicense < ActiveTriples::Resource
+    class DummyLicense
+      include ActiveTriples::Entity
       property :title, :predicate => RDF::DC.title
     end
 
-    class DummyResource < ActiveTriples::Resource
+    class DummyResource
+      include ActiveTriples::Entity
       configure :type => RDF::URI('http://example.org/SomeClass')
       property :license, :predicate => RDF::DC.license, :class_name => DummyLicense
       property :title, :predicate => RDF::DC.title
@@ -200,7 +202,8 @@ describe ActiveTriples::Resource do
     subject {DummyResourceWithBaseURI.new('1')}
 
     before do
-      class DummyResourceWithBaseURI < ActiveTriples::Resource
+      class DummyResourceWithBaseURI
+        include ActiveTriples::Entity
         configure :base_uri => "http://example.org",
                   :type => RDF::URI("http://example.org/SomeClass"),
                   :repository => :default
@@ -246,7 +249,8 @@ describe ActiveTriples::Resource do
     subject {DummyResourceWithBaseURI.new('11')}
 
     before do
-      class DummyResourceWithBaseURI < ActiveTriples::Resource
+      class DummyResourceWithBaseURI
+        include ActiveTriples::Entity
         configure :base_uri => "http://example.org",
                   :type => RDF::URI("http://example.org/SomeClass"),
                   :repository => :default
@@ -403,7 +407,7 @@ describe ActiveTriples::Resource do
       end
 
       it 'should return generic Resources' do
-        expect(subject.attributes[RDF::DC.relation.to_s].first).to be_a ActiveTriples::Resource
+        expect(subject.attributes[RDF::DC.relation.to_s].first).to be_a ActiveTriples::Entity
       end
 
       it 'should build deep data for Resources' do
@@ -527,7 +531,7 @@ describe ActiveTriples::Resource do
         subject.set_value(RDF::DC.title, RDF::URI("http://opaquenamespace.org/jokes/1"))
       end
       it "should return a resource" do
-        expect(subject.title.first).to be_kind_of(ActiveTriples::Resource)
+        expect(subject.title.first).to be_kind_of(ActiveTriples::Entity)
       end
       context "and it's configured to not cast" do
         before do
@@ -698,14 +702,16 @@ describe ActiveTriples::Resource do
 
   describe 'big complex graphs' do
     before do
-      class DummyPerson < ActiveTriples::Resource
+      class DummyPerson
+        include ActiveTriples::Entity
         configure :type => RDF::URI('http://example.org/Person')
         property :foaf_name, :predicate => RDF::FOAF.name
         property :publications, :predicate => RDF::FOAF.publications, :class_name => 'DummyDocument'
         property :knows, :predicate => RDF::FOAF.knows, :class_name => DummyPerson
       end
 
-      class DummyDocument < ActiveTriples::Resource
+      class DummyDocument
+        include ActiveTriples::Entity
         configure :type => RDF::URI('http://example.org/Document')
         property :title, :predicate => RDF::DC.title
         property :creator, :predicate => RDF::DC.creator, :class_name => 'DummyPerson'
@@ -772,7 +778,8 @@ END
   describe "callbacks" do
     describe ".before_persist" do
       before do
-        class DummyResource < ActiveTriples::Resource
+        class DummyResource
+          include ActiveTriples::Entity
           def bla
             self.title = "test"
           end
